Add files via upload
Enter Registered Email: The user enters their registered email address in the provided text field.
Request Password Reset: After clicking the "Reset Password" button, the app sends a request to Firebase Authentication to send a password reset email.
Validation and Feedback: The email is validated, and if the email is valid and registered, Firebase sends the password reset link. If there’s an error, the app provides a descriptive message to guide the user.
User Feedback: A confirmation message is displayed in the form of a SnackBar, notifying the user of the outcome.